% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArrangeGgplots.R
\name{arrangeGgplots}
\alias{arrangeGgplots}
\title{Arrange Multiple ggplot Objects with Shared Axes and Labels}
\usage{
arrangeGgplots(ggplotList, layoutInstruction, shareX = FALSE, shareY = FALSE)
}
\arguments{
\item{ggplotList}{A list of ggplot objects to be arranged.}

\item{layoutInstruction}{A string representing the layout instruction for the \code{patchwork::wrap_plots} function.}

\item{shareX}{Logical; if \code{TRUE}, the X axes will be shared across plots. Defaults to \code{FALSE}.}

\item{shareY}{Logical; if \code{TRUE}, the Y axes will be shared across plots. Defaults to \code{FALSE}.}
}
\value{
A ggplot object representing the combined layout of the input plots.
}
\description{
This function arranges multiple ggplot objects into a single plot layout,
with options to share X and Y axes as well as labels.
It uses the \code{patchwork} package to create a composite plot.
}
\examples{
# Example usage:
# arrangeGgplots(list(ggplot1, ggplot2), layoutInstruction = "1|2", shareX = TRUE)
}
