% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateBoxPlot.R
\name{createBoxPlot}
\alias{createBoxPlot}
\title{Create a Box Plot}
\usage{
createBoxPlot(df, title, yLabel, xLabel)
}
\arguments{
\item{df}{A data frame containing the required columns for plotting.}

\item{title}{The title of the plot.}

\item{yLabel}{The label for the y-axis.}

\item{xLabel}{The label for the x-axis.}
}
\value{
A ggplot object representing the box plot.
}
\description{
This function creates a box plot using ggplot2, given a specific data frame structure.
It requires certain columns in the data frame and adds annotations for statistical
measures like mean and standard deviation.
}
\examples{
# Example usage:
# createBoxPlot(df = yourDataFrame, title = "Your Title", yLabel = "Y-axis Label", xLabel = "X-axis Label")
}
