% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateTrellis.R
\name{createTrellis}
\alias{createTrellis}
\title{Create a Combined Trellis Plot from a List of ggplot Objects}
\usage{
createTrellis(ggplotArray, maxColumns)
}
\arguments{
\item{ggplotArray}{A list of ggplot objects to be combined into a trellis plot.}

\item{maxColumns}{The maximum number of columns to use in the trellis layout.}
}
\value{
A combined ggplot object representing the trellis plot.
}
\description{
This function takes an array of ggplot objects and combines them into a single trellis plot.
The plots are arranged in a grid with a specified maximum number of columns.
}
\examples{
# Assuming you have a list of ggplot objects named myPlots
combinedPlot <- createTrellis(myPlots, 2)
print(combinedPlot)
}
