% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OhdsiPlots.R
\docType{package}
\name{OhdsiPlots-package}
\alias{OhdsiPlots}
\alias{OhdsiPlots-package}
\title{OhdsiPlots: An R package to compute temporal relationships between persons in two or more instantiated cohorts.}
\description{
An R package that takes output from OHDSI Results Model and provides plots for publications.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ohdsi.github.io/OhdsiPlots/}
  \item \url{https://github.com/OHDSI/OhdsiPlots}
  \item Report bugs at \url{https://github.com/OHDSI/OhdsiPlots/issues}
}

}
\author{
\strong{Maintainer}: Gowtham Rao \email{rao@ohdsi.org}

Other contributors:
\itemize{
  \item Observational Health Data Science and Informatics [copyright holder]
}

}
\keyword{internal}
