% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotCompareStandardizedDifference.R
\name{plotCompareStandardizedDifference}
\alias{plotCompareStandardizedDifference}
\title{Plot Temporal Compare Standardized Difference}
\usage{
plotCompareStandardizedDifference(
  df,
  domain = "all",
  threshold = 1000,
  thresholdValue = 0.01,
  facetOn = NULL,
  plotTitle = "Comparison of Standardized Differences",
  xAxisLabel = "Prevalence in Target Cohort",
  yAxisLabel = "Prevalence in Comparator Cohort"
)
}
\arguments{
\item{df}{Data frame containing the data to plot.}

\item{domain}{Domain to filter on, defaults to "all".}

\item{threshold}{Threshold for row count filtering, defaults to 1000.}

\item{thresholdValue}{Value threshold for filtering data, defaults to 0.01.}

\item{facetOn}{(Optional) Column name to use for faceting.}

\item{plotTitle}{Title of the plot, defaults to "Temporal Comparison of Standardized Differences".}

\item{xAxisLabel}{Label for the X-axis, defaults to "Prevalence in Target Cohort".}

\item{yAxisLabel}{Label for the Y-axis, defaults to "Prevalence in Comparator Cohort".}
}
\value{
A ggplot object representing the plot.
}
\description{
This function generates a static plot to compare standardized differences over time.
It allows for optional faceting based on a specified column and includes various parameter checks.
}
\examples{
# Example usage:
plot <- plotCompareStandardizedDifference(df, domain = "all")#'
}
