% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotTrendCurve.R
\name{plotTrendCurve}
\alias{plotTrendCurve}
\title{Plot a Trend Curve}
\usage{
plotTrendCurve(
  df,
  dateColumn,
  smoothingMethod = "auto",
  showStandardError = TRUE,
  outputFilename = NULL,
  plotTitle = "Trend Curve",
  xAxisLabel = "Date",
  yAxisLabel = "Volume",
  yScaleTransform = "identity",
  ggplotTheme = ggplot2::theme_minimal(),
  dateLabelDf = NULL,
  cumulative = FALSE,
  showRawValues = FALSE,
  showCumulativeValues = FALSE,
  cumulativeYPosition = -10,
  cumulativeJitter = 3,
  dateRounding = "none"
)
}
\arguments{
\item{df}{A tibble containing the data to be plotted.}

\item{dateColumn}{A string specifying the column name in \code{df} that contains date values.}

\item{smoothingMethod}{A string indicating the smoothing method to be used. "auto" by default.}

\item{showStandardError}{Logical indicating whether to show standard error in the plot. Defaults to TRUE.}

\item{outputFilename}{A string specifying the path to save the plot image. If NULL, the plot is not saved. Defaults to NULL.}

\item{plotTitle}{A string for the plot title. Defaults to "Trend Curve".}

\item{xAxisLabel}{A string for the x-axis label. Defaults to "Date".}

\item{yAxisLabel}{A string for the y-axis label. Defaults to "Volume".}

\item{yScaleTransform}{A string specifying the transformation for the y-axis scale. Defaults to "identity".}

\item{ggplotTheme}{A \code{ggplot2} theme object. Defaults to \code{theme_minimal()}.}

\item{dateLabelDf}{A tibble with columns 'date' and 'label' for annotating specific dates. NULL by default.}

\item{cumulative}{Logical indicating if the plot should display cumulative values. Defaults to FALSE.}

\item{showRawValues}{Logical indicating if raw data points should be displayed on the plot. Defaults to FALSE.}

\item{showCumulativeValues}{Logical indicating if cumulative values should be annotated on the plot. Defaults to FALSE.}

\item{cumulativeYPosition}{Numeric indicating the Y-axis position for cumulative value annotations. Defaults to -10.}

\item{cumulativeJitter}{Numeric indicating the amount of jitter to apply to cumulative annotations to avoid overlap. Defaults to 3.}

\item{dateRounding}{A string specifying the unit for rounding dates. "none" by default.}
}
\value{
None
}
\description{
This function creates a trend curve plot from a given data tibble. It allows for various customizations including smoothing methods, showing standard error, cumulative data, raw values, and more.
}
