% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotWeighted.R
\name{plotWeighted}
\alias{plotWeighted}
\title{Plot Weighted Data}
\usage{
plotWeighted(
  df,
  x = "value",
  y = "weight",
  mainTitle = "Weighted Plot",
  xLabel = "Values",
  yLabel = "Weights",
  plotColor = "blue",
  plotType = "histogram",
  binwidth = NULL,
  useScinot = FALSE,
  plotTheme = "theme_minimal"
)
}
\arguments{
\item{df}{A data.frame or a vector. If a vector is provided, it's converted into a data.frame with default weights of 1.}

\item{x}{The name of the variable to be plotted on the x-axis. If data is a vector, this defaults to "value".}

\item{y}{The name of the variable representing weights. If data is a vector, this defaults to "weight".}

\item{mainTitle}{The main title of the plot.}

\item{xLabel}{The label for the x-axis.}

\item{yLabel}{The label for the y-axis.}

\item{plotColor}{The color to be used for plotting.}

\item{plotType}{The type of plot to create, either "histogram" or "bar".}

\item{binwidth}{The bin width for the histogram. If NULL, it is automatically determined.}

\item{useScinot}{A logical indicating whether to use scientific notation or not.}

\item{plotTheme}{The ggplot2 theme to be applied to the plot.}
}
\value{
A ggplot object representing the weighted plot.
}
\description{
Creates a weighted plot (histogram or bar plot) based on the provided data and weights.
The function allows customization of various plot aspects such as title, labels, color, and plot type.
}
\examples{
plotWeighted(mtcars, x = "mpg", y = "wt", mainTitle = "Car Weight by MPG", plotType = "bar")
}
