% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RenderTrendGraph.R
\name{renderTrendGraph}
\alias{renderTrendGraph}
\title{Render Trend Graph}
\usage{
renderTrendGraph(
  df,
  xAxisLabel = "Date",
  yAxisLabel = "Volume",
  yScaleTransform = "identity",
  ggplotTheme = ggplot2::theme_minimal(),
  showRawValues = FALSE,
  groupBy = NULL,
  smoothLines = FALSE,
  labelGroup = FALSE
)
}
\arguments{
\item{df}{A data frame containing the data to be plotted. Must contain columns 'date' and 'count'.}

\item{xAxisLabel}{A string specifying the label for the x-axis. Default is "Date".}

\item{yAxisLabel}{A string specifying the label for the y-axis. Default is "Volume".}

\item{yScaleTransform}{A string specifying the transformation for the y-axis scale. Can be "identity", "log", "sqrt", etc.}

\item{ggplotTheme}{A ggplot2 theme object. Used to customize the appearance of the graph.}

\item{showRawValues}{Logical, whether to show raw values as points on the graph.}

\item{groupBy}{A string specifying the column name in 'df' to group the data by. NULL means no grouping.}

\item{smoothLines}{Logical, whether to use smooth lines instead of regular lines.}

\item{labelGroup}{Logical, whether to label the groups if grouping is used.}
}
\description{
This function creates a trend graph based on the given data frame. It allows customization of various aspects
of the graph such as axis labels, transformations, and grouping.
}
