% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OhdsiPlots.R
\docType{package}
\name{OhdsiPlots-package}
\alias{OhdsiPlots}
\alias{OhdsiPlots-package}
\title{OhdsiPlots: Another R package for plots.}
\description{
Another R package for plots.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ohdsi.github.io/OhdsiPlots/}
  \item \url{https://github.com/OHDSI/OhdsiPlots}
  \item Report bugs at \url{https://github.com/OHDSI/OhdsiPlots/issues?}
}

}
\author{
\strong{Maintainer}: Gowtham Rao \email{rao@ohdsi.org}

}
\keyword{internal}
